requirements.txt
Flask==2.0.1
python-dotenv==0.18.0
celery==4.4.1
gunicorn
psycopg2-binary==2.9.1
requests
Flask-SQLAlchemy
